make more generic AbstractChoiceDialog
